backend/pyproject.toml
[project]
name = "northlink-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt<4.1",
    "cryptography>=46.0.5",
    "fastapi>=0.134.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "sentry-sdk>=2.53.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]
